Enable json marshalling of Spans in IndexScan. Fixes MB-13597.
Fix json marshalling of planner.Spans and add json unmarshalling
of planner.Spans. The former fixes use of named/positional parameters
in EXPLAIN and PREPARE statements that involve a secondary index.
The latter fixes EXECUTE of statements that involve a secondary
index. Also adds json marshalling for the child operator of an
Authorize operator.

Change-Id: Ie9b3e6865bbc7752af7a3b2fd436313fbcc9f340
Reviewed-on: http://review.couchbase.org/47933
Reviewed-by: Manik Taneja <manik@couchbase.com>
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Gerald Sangudi <gerald@couchbase.com>