Initial connection between cbq-engine and the schema inferencing code
Ammended check-in. Needed to fix error due to outdated cb library (go get is not infallible).

Also call NewDefaultSchemaInferencer only once.

Also refer to new location for cbq-gui.

Finally, change the build so that in Enterprise mode it pulls in inference code from couchbase/cbq-gui
(which is a private repository), but in Community builds it will instead just use dummy code that will
return an "not implemented: INFER" error.

Change-Id: Ieceaedb0e356a639f9b3104b7ab35e6fec0bafe1
Reviewed-on: http://review.couchbase.org/58959
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Reviewed-by: Eben Haber <eben@couchbase.com>
Tested-by: Eben Haber <eben@couchbase.com>