MB-15846 Allow blank query parameters.
The users want to be able to run queries like this:

curl http://localhost:8093/query -d "statement=select * from default; "

Note the semicolon and the space after it. Unfortunately
the semicolon and space are interpreted as a new parameter
named with spaces. Our code never sees the (superfluous)
semicolon in this case.

To accommodate this case, we allow unknown parameters that are
purely whitespace.

Change-Id: I1c24b7c7f22f8fc65ce6e4636b2f78958f87402b
Reviewed-on: http://review.couchbase.org/59974
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Johan Larson <johan.larson@couchbase.com>