MB-17263 Use a single encoded value rather than an array.
The key of a primary index of view type is an encoded single value,
rather than an encoded array of values, as in the secondary indexes
of view type. Accordingly, encode the search keys used for these
primary indexes as single values, not arrays of values.

The code panics if the incoming value is compound. Such a value
should never have been sent by the planner anyway.

Change-Id: I5d93c8a2c08d3ed6f8952e14fce1e7a5e4fc19d5
Reviewed-on: http://review.couchbase.org/63359
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Johan Larson <johan.larson@couchbase.com>