MB-19364 Fix count
The Count() of system:keyspaces was reversed, causing the code
to count the number of memcached buckets rather than the number
of ordinary couchbase buckets. I didn't see the problem because
I was testing with exactly one of each kind, so I saw the number
I wanted, and it wasn't 2. My changes also broke a test, which
had scrolled off my screen. Mea culpa.

Change-Id: I07e8e53fbab1d3dc2bc0037dfbed0c974d461258
Reviewed-on: http://review.couchbase.org/63536
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Johan Larson <johan.larson@couchbase.com>