MB-20574 Improve error message for unsupported view index.
This change does not fix the underlying problem, but I have been asked for a quick patch rather
that a full resolution of the problem, since view indexes are not used much.

This change improves the error message to explain what might be going wrong,
and makes the problematic view definition available in the log. This makes the
line number presented in the original error message useful for debugging.

The new error message conveyed to the user is:

   "Failed to create view idx2 - cause: Unable to store the view definition.
   Not all index target expressions are supported. Check whether the JavaScript of the view definition is valid.
   The map function has been output to query.log.
   Original error: error installing view: 400 Bad Request / {\"error\":\"invalid_design_document\",
   \"reason\":\"Syntax error in the map function of the view `idx2': SyntaxError: Unexpected token ( (line 52:37)\"}\n"

Change-Id: Icf5324424290fd8db199460b34a037eb9fbddab4
Reviewed-on: http://review.couchbase.org/67089
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Johan Larson <johan.larson@couchbase.com>