MB-20967 clean up fields in clustered monitoring
The accounting REST interface and system keyspaces name  monitoring
fields differently, and in cases have different fields.
This so far has not been changed to avoid issues in QA, but needs to
be changed before 4.6 because otherwise clustered query monitoring would
return different types of documents depending if the entry is local or
remote, which would be pretty poor.

I have checked tuqquery and don't believe QA would be impacted.
I have to check with the Query WorkBench team to make sure they haven't
started work and would  be impacted.

The bulk of the work is to standardise element names to starting with
lowercase, eliminating various spurious dots in element names,
making sure that names are consistent in between keyspaces (ie the same
thing named in two different ways in two different keyspaces) and making
that local and remote entries return the same fields.
For prepareds, remote entries returned a full plan.
This is now being set as an attachment called 'plan'.
(select meta().plan from system:prepareds works).
I plan to extend this to profiling when I get to code taht.

Change-Id: I8e19ea4cfb85d2077d59ab35fc9ccc0ebd6c7dc9
Reviewed-on: http://review.couchbase.org/67885
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>