MB-20183 explain text is truncated
This tries to make the token offset reliably part of the Lexer type
in a way that's not version specific.
The offset is inserted at build time in the generated parser by AWK
searching for the NEX_END_OF_LEXER_STRUCT tag.
Not the cleanest of methods, but apparently the only supported one.

Change-Id: I405ea0c4d3d51250c13c7d5fb9e428a514f6d4f4
Reviewed-on: http://review.couchbase.org/68165
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>