MB-20979 improve http chunked transfer rate
service_response.go uses two writers to return results to the client:
initially a buffered writer, and later a direct writer.
The idea being, we'll buffer up to a size and  if the response is less
than that size, we'll send it all in one go.
If not we initiate chunked transfer.
The slight problem is that the direct writer flushes every string, and 
the act of flushing makes the http server produce chunks, so with this 
dual scenario we produce very short chunks.

Changed to use a single buffered writer. If we can write all within the
buffer  size, fine, if not we still buffer, but flush every 100ms, so
that we still transfer large chunks, and keep the client busy given the
short flush intervals.

With this change, select * from `travel-sample` on my machine completes
in 5.1 secs vs 6.2 w/o for a result set size of 100MB.

Change-Id: I6cb5e924847cd62c680c7c5c5383ac11fcd40aaf
Reviewed-on: http://review.couchbase.org/68226
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>