MB-21028 clustering can't determine node name in a single node cluster
When nodes are all named 127.0.0.1, ns_server is smart enough to return
actual non local ip addresses.
Not so when there's a single node and ns_server is accessed locally.
Http clustering code rejects 127.0.0.1 as a proper node name, which
means that on single nodes, the n1ql service thinks it is not part of a
cluster.
If 127.0.0.1 should come our way - fix it.

Change-Id: Id59d6ae3abc0fe203cd936813a3f44c05da9843d
Reviewed-on: http://review.couchbase.org/68332
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>