MB-21290 Support authorization by cbauth token rather than username/password.
This implementation has the unfortunate effect of sending the
HTTP request all the way down into the datastore component.
This is because the cbauth method we have to call only accepts
the full request.

The alternatives would be to do the call earlier in the execution chain,
but that would mean we are calling cbauth in two places in the code,
or rewriting the cbauth interface to separate the extraction of the
authorization data and their verification.

Neither option seemed obviously superior to what is implemented here.

Change-Id: If6a66cb6bdc1642db0b2afed4da51cb8e68072b0
Reviewed-on: http://review.couchbase.org/68632
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Johan Larson <johan.larson@couchbase.com>