MB-21231 avoid quoting hell
As explained in the defect comments, the underying issue is that object
field names get quotes added to them in the stringer because string
values are marshalled.
As we marshal over and over objects, field names have extra quotes added.
In this case, the index definition has quotes added in the where clause
when passed to the indexer.
When sarging it, the where clause in the query has one extra set of
quotes added, but the index condition, starting with an extra set, ends
up with two extra sets of quotes, and sargableIndexes() deems the two
conditions not equivalent.

The fix is to stop quoting hell.
We can't force dynamic object construction in the parser to only accept
string literals as field names (it breaks tests), and we can't prevent
stringValue.String() from marshalling, as it would break so many things,
so the only way to fix this is to make sure that  the stringer only
marshals dynamic object field names if required.

The upshots of this fix are that
1- all index with dynamic objects have to be recreated
2- any dynamic object appearing in prepare or explain outputs (eg
conditions, filters, projections, etc) now don't have the extra quotes
attached (good - as they are now as they should be), so any test that
relies o the extra quotes will have to be changed.

This may slow down  QA in completing  watson testing.
There may also be other unforseen side effects.
Still - I was asked for a fix, here's one....

Change-Id: I6b0764bc47650eb59458ae48c2d35f91ad4e6414
Reviewed-on: http://review.couchbase.org/68756
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>
(cherry picked from commit 8b2a84dcd3cda814f46c251b0cccf7d38adb8d59)
Reviewed-on: http://review.couchbase.org/68814