MB-21527 race condition on request objects
The underlying issue is that access to request objects for marshalling
purposes is done outside of a lock. We should have had prepareds and
active_requests match the implementation of completed_requests (which
uses a multi lock cache and a locked access Get() method): this was
scheduled to happen  together with MB-18290, but we have not done this
in Watson due to the complexity of the fix.
There's enough of an initial code change here to prevent the crashes,
but this likely has performance implications.

Change-Id: I377046926c00b9fb4255b5c340887ac1dd1e29a1
Reviewed-on: http://review.couchbase.org/69924
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>