MB-21723 align AddPhaseTimes with AddPhaseCounts / Operators
This looks like a major change, but in actual fact most of the changes
are switching a parameter from a string to a Phase.

We are switching to AddPhaseTime() to the same format as AddPhaseCounts()
etc for maintainability and performance as per profiling design spec.

I'll wait for Sitaram to  merge his change before merging, as they 
conflict...

Change-Id: I83ca37b15cc51e0ae5cec3158b9276078614fde2
Reviewed-on: http://review.couchbase.org/70061
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>