MB-20969 create execution layer marshaller for profiling epic
This looks massive, and it is - but the changes are almost mechanical.
The idea is to create marshalling infrastructure in plan that execution
can use for marshalling so that it matches plan marshalling.
The objective here is that any change in plan are automatically
reflected in execution marshalling with the only extra task needed being
to handle children, should new children be added to an operator.

Other activities have entailed moving time accrual from plan to
execution. The behaviour so far coded matches exactly what was already
implemented for request debugging. Proper query profiling work will be
done in other stories. This is just groundwork.

One drawback is that I have had to - and I stress, temporarily - remove
pooling from 4 execution operators. This is because pooling ditches
children before we get to marshal them, and essentially we get truncated
timing at the first sequence (all you get is "~children": null).

This will be corrected ASAP, but as part of a larger effort that we need
to discuss.
Essentially we will need a execution.Close() method that we can call on
the root operator after we have marshalled (the are other options too, eg
marshalling can automatically close and we only explicitly close if we
don't have to marshal) so that all children can get closed.

Having a Close() method will allow us to do funky stuff like pooling
operators as well as []Operators and channels, which may give us some
extra performance boost by saving yet more GC activity.

Change-Id: I27e8b3373c296d0c68c3e2e3ea22341430afef82
Reviewed-on: http://review.couchbase.org/70370
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>