MB-21920 add a Done method in execution to reinstate pooling
This adds a Done() method to the execution layer to signal when
it's fine to be done with execution resources.
This is required to reenable []operator pooling.
Several notes
1- the method is called Done() and not Close() because Close() was
already taken (Channel operator closing)
2- although it was initially suggested that MarshalJSON() might
implictly release resources (so as not to have to traverse the operator
tree twice), it turns out - reading the release specs, that we may
marshal during request execution (as system:active_requests is queries)
and implictly releasing resources will break havoc
3- this is only the initial effort needed to get the current implementation
of pooling going again. Further pooling (eg Operators, as well as
[]Operators) still has to be designed and discussed, and will be dealt
with separately

Change-Id: I89b42b8a9152ddc8dba6d7cf0e6889056fc9f258
Reviewed-on: http://review.couchbase.org/70727
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>