MB-20971 add command line + REST options for profiling MB-19762 include parameters in metrics
This adds command line options and /admin/settings endpoint for profile
and controls in the output document.
It also implements profiles and controls (for now the parameters) in the
document.
Implementing profile and metrics in the monitoring  keyspaces will require
a bit of extra work (including RBAC) because of distrbuted monitoring:
we need to make a little extra effort to get profiles and controls from
remote nodes...

Change-Id: I7a4a7a836db4e9ba374fa83c7b00a8f574811fd4
Reviewed-on: http://review.couchbase.org/70913
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>