MB-21011 EXPLAIN PREPARE mistakenly insert statement in cache
While we have to cleanup the statements that PREPARE and EXPLAIN accept
(so as to avoid needless cache clogging and prepare nesting in the
results document), this is part of a separate defect.
What we deal here is the fact that a prepared statement is added to the
cache while planning, not during execution, which is incorrect, and leads
to prepared statements being added when it was not intended.
Moved AddPrepared call to execution - also copied *plan.Prepared in 
execution/prepare.go so as to avoid the double penalty hit of marshalling
the plan into a value in the planner, and unmarshalling it back in
execution.

Change-Id: I316aa6bbd224bf4ad57b45ab61c3a350b162487c
Reviewed-on: http://review.couchbase.org/71494
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>