MB-20970 complete timings MB-22354 standardise explain children names
This completes profiling bar one thing: currently the parallel operator
only provides timings for one child.
This is so that explain and timings output actually match (otherwise
timings would have to have a ~children array, a mismatch with explain).

The other thing to not is that KV based operators (index scans and joins,
mostly) only shown #itemsOut, but no #itemsIn (which would be coming)
from the KV, so that they are actually identifiable as producers.

Lastly - for now I have fixed "~child" and "~children" in plan, which
I know (as I have checked) it won't affect tuqquery.
We'll do "first", "second", "scan" and "scans" later on.

Change-Id: I65916fc7846f820a48981cac7969743d02ef5df9
Reviewed-on: http://review.couchbase.org/72128
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>