MB-20970 profiling should help spotting looping or deadlocked operators
Last friday Keshav and I were debugging a stuck request, which as it turns
out had an operator stuck on a sendItem().
Since times are only accrued as phase switches occur, when an operator is
stuck, times don't really change, and there's no way to determine what the
state of each operator is.

With this change, the marhaller will accrue the time since the last
switch and display the current phase if the operator has not yet
completed work.

In short, for active requests, the times will be up to now, while, for
completed requests and profiling, since the operators have completed,
the code will not be exercised.
In the event of a runaway operator - completed_requests will show
operators that still accrue time, which again is a useful debugging tool.

The infrastructure also keeps track of the number of context switches,
so it's pretty easy to tell what isn't moving, as the context switches
won't increase.


Change-Id: I07ba765b59b25ee2ad0f21a522fded8f1c89d696
Reviewed-on: http://review.couchbase.org/72362
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>