MB-20773 add secondary index on node for monitoring keyspaces
the secondary index on the node field is only enabled if the node is
found to be part of a cluster.
it only supports seeks because the SystemRemote interface does not
currently report the topology - just uses it.

This needs to be functionally tested properly as right now  the planner
does not do scanner for system keyspaces because of MB-19445/MB-19475
which have the nice effect of crashing the service.

I'll validate this code once I have fixed those - will merge but keep
the MB open.

Note that the scan method for these keyspaces does not do seeks on the
primary index (and reverts to a primary scan) because remote scans are
always primary and do not currently support seeks.
This is no different than what it has always been.

Change-Id: I1617f6d82698720a63da305683e868dd92ac6259
Reviewed-on: http://review.couchbase.org/72470
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>