MB-22205 filter INDEXES and KEYSPACES based on the submitted credentials
The is a bit of weird magic here that I hope Marco can help me understand.
The code the does the permission filtering accesses _DATASTORE
to get the current datastore. To make the tests pass I therefore have to
make sure it's set, and I do it in test/filestore/json.go and
test/multistore/json.go. No problem so far.

Unfortunately that breaks the tests under system/testcs_system
that deal with system:completed_requests. I can make the tests
pass by unsetting _DATASTORE again in this test. But I'm not sure
why this is happening. Marco, any ideas?

Change-Id: I24470e9f52422b2235d4a8b25c950bd8ca1aad60
Reviewed-on: http://review.couchbase.org/72533
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Johan Larson <johan.larson@couchbase.com>