MB-20773 make sure that planner selects #nodes index It took me the best part of a day to debug this by instrumentation. The planner wasn't selecting the #nodes index because the system indexer would only ever return the primary index.
Given that the system indexer is static, we could improve in system.go
by allocating the slices for indexes, primary indexes, names, etc when
the keyspaces are created - this would save some GC work, but for
now this will do.

What's left now is to fix MB-22588 (the clustering package misses nodes)
and to make sure that the planner doesn't use the secondary index for
sorts.

Change-Id: I32da54dbbdf3b40992a7c0eb94867a91fe8faa49
Reviewed-on: http://review.couchbase.org/73015
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>