MB-21085 allow requesting multiple privileges per keyspace
A key part of the RBAC design is that running a statement requires
two types of permissions: you need permission to run the statement
type itself for the buckets, and you need permission to access the
actual data in the bucket. We therefore require multiple privileges
per bucket. The current codebase stores privileges as a map
from bucket to privileges, which does not allow this.

In this checkin, I replace the Privileges map by a struct holding
a list of target+privilege pairs. (in privilege.go) This requires
far-reaching but minor changes to function interfaces.

There is also code introducing statement types for SELECT, INSERT,
DELETE, and UPDATE, but pending our discussions with Don, these
are automatically granted and are hence no-ops.

Change-Id: I5a837c1cb7c470434a1c1dea548c0872e5eb9a73
Reviewed-on: http://review.couchbase.org/73461
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Johan Larson <johan.larson@couchbase.com>