MB-22677 backfills not deleted
There is indeed a race condition in between RunOnce() and Done()
Added a WaitGroup and used whenever operators have to wait for children.

Have kept this purposly simple and have specific operators that need to
be mindful about timings when executing Done() explicitly signalling
that they are active / inactive, rather than having more generalized
infrastructure (like calling wg.Add(1) when we switch out of _NOTIME
and wg.Done() when we walk into to, or wg.Add(1) in newBase() and
wg.Done() in notify()) because that could actually end up generating
negative waitgroups, if we are not careful!


Change-Id: I6c8be6c1dcc887f440eccfdd5f029b38309d61d2
Reviewed-on: http://review.couchbase.org/73625
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>