MB-21099 Enable GRANT ROLE through the DO_ADMIN function.
This required some rejigging of interfaces to avoid a circular dependency
between /expression and /datastore; the User and Role structs
moved to /expression, and execution.Context got a couple of new functions
so the expression code doesn't have to directly refer to /datastore.

Note that the same code is now executed by both the statement version
of GRANT ROLE and the DO_ADMIN version. The behavior has changed slightly
as a result. Where before the code would churn through all the users needing update,
it now stops after the first error since the DO_ADMIN version can only accommodate
a single error.

Change-Id: Idd0ae719c5d1e13a7ea57cdc2a34eac77d0cfe60
Reviewed-on: http://review.couchbase.org/73844
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Johan Larson <johan.larson@couchbase.com>