MB-23032 fix goroutine leak in scanner Summary - nex has always had a scanner leak issue, which has been recently fixed. The fix is not complete, so I have rewritten it.
After discussion with Ceej, we have a fixed nex in couchbaselabs, with
the fix, and I plan to have my fix merged in the original repository.
We can keep couchbaselabs/nex as a reliable source for nex.

Have changed the build script to go get couchbaselabs/nex if a new
scanner is required, hence we no longer need to have nex in our machines,
and have changed n1ql.go:doParser() to Stop() the scanner threads after
parsing, even after a recover.

No change is required to the build process, and no change is required
to anyone's machine, apart from ditching your copy of nex, if you so
wish.

Change-Id: I7bd04ce9ce7a634db28fb5d84ed25eb4c24f6ece
Reviewed-on: http://review.couchbase.org/74404
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>