MB-19097 add infrastructure to have multiple completed_requests qualifiers
We have had one recent CBSE where it would have been very nice to be able
to log requests which were generating syntax errors (just because the
customer was having tens of thousands of syntax errors by the hour).
With this, we add infrastructure in system:completed_requests to be able
to apply filters to the request and decide if logging is necessary based
on filters different from execution time.

For now the actual functionality is unchanged.

In order to add more filters, the steps required are
- create a new qualifier
- add a command ine switch to set the filter
- add /admin/settings checkers and setters to set / unset / display
  the new filter.

Change-Id: I2016a0fe8975ed60bdfc2f6102d87926d8ff3fbf
Reviewed-on: http://review.couchbase.org/75220
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>