MB-22110 syntax errors not detected
The underlying issue is that nex by default just skips what it does not
match, whereby lexers are expected to tokenize everything bar comments
and whitespace.
So added a specific match at the end of the lexer to match the current
character and pass it to the parser.
The parser will then throw an error, unless it knows how to handle
that character (our parser will reject out of hand).

Note that it is perfectly fine (and indeed an old school technique) to
pass ascii values to yacc, as all the token values generated by yacc
are by default > 256 specifically to handle this.

With this, we produce:

cbq> select 1, 2, ##3;
{
    "requestID": "d58b48ed-623c-4c1c-ae83-ffe738cf69d3",
    "errors": [
        {
            "code": 3000,
            "msg": "syntax error - at #"
        }
    ],
 ...
}

Also checked that explain and prepare still extract statement text
properly:

cbq> explain    
   > 
   > select
   > 1
   > ,2,3;
{
    "requestID": "7d726637-1a0f-47b9-a657-004651c3d3e3",
    "signature": "json",
    "results": [
        {
            "plan":
            ...
            "text": "select 1 ,2,3;"
        }
    ],
...
}

Change-Id: Iaebee6ecc478ab917fe66b0ec7c20782f31a1c02
Reviewed-on: http://review.couchbase.org/75338
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>