MB-23103 tag timings with version MB-23378 no meta().plan for remote nodes
We are tagging the root operator of the timings tree with a version, so
every place with output timings (profile section of output, meta().plan
in system:completed_requests, system:active_requests), we also get both
the n1ql and the datastore version, which is useful for diagnostic
purposes (the N1QL version is the main indicator for planner issues, the
datastore version is there just to detect if there's a mismatch between
n1ql service and cluster, which could be useful for unexpected behaviour).

The versions get output at the end of the timings, so to be more prominent.

cbq> select *, meta().plan from system:completed_requests;
{
    "requestID": "98d80ed3-def7-4e11-95d9-ddc73a4fc8a9",
    "signature": {
        "*": "*",
        "plan": "json"
    },
    "results": [
        {
            "completed_requests": {
                "elapsedTime": "5.659152712s",
                "errorCount": 0,
                "node": "192.168.42.72",
...
            },
            "plan": {
                "#operator": "Sequence",
                "#stats": {
                    "#phaseSwitches": 2,
                    "execTime": "1.975µs",
                    "kernTime": "5.629892022s"
                },
                "~children": 
...
                "~versions": [
                    "1.7.0-N1QL",
                    "5.0.0-2261-enterprise"
                ]
            }
        }
    ],
    "status": "success",
    "metrics": {
        "elapsedTime": "60.049335ms",
        "executionTime": "60.03331ms",
        "resultCount": 1,
        "resultSize": 7116
    }
}

I will get in touch with Ceej to study a way to add the build number
to the N1QL version.

MB-23378, by code inspection the timings subdocument wasn't being
marshalled in the POST /admin/[completed|active]_request endpoints.

Change-Id: Ibbcc996c67f6122128da45d623360dc222f655a0
Reviewed-on: http://review.couchbase.org/75354
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>