MB-22709 improvements to N1QL clustering code
MB-22709 lists just about anything that can go wrong with N1QL
clustering including multiple heads, inability to determine our own
name, inability to determine down nodes, etc.

This is a first improvement on that front, whereby we really try to
identify the node name, taking into account NIC failures and multiple
NICs (and a mismatch in between what ns_server thinks our IP address is
and what we think).
We do that by using all the NICs that have an IPV4 address for the search
and checking the the NIC state after we find the one that applies.

This is related to MB-23420 whereby when a NIC goes down, we don't
find our original name in the list of N1QL node and report monitoring
entries twice (once local, and once remote because we fail to identify
a remote node as ourselves).

There's much more to do, but at least this addresses IP address issues.


Change-Id: I1b55763ebf42de011509cc5b463edbddc9060999
Reviewed-on: http://review.couchbase.org/75510
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>