MB-19601 cbq-engine polling for design docs to port 8091 at a rate of 1100 requests per second on otherwise idle cluster
This is a defensive fix based on this hint in the log

_time=2016-05-11T01:46:20.720-04:00 _level=INFO _msg=Got new configuration for bucket default 
_time=2016-05-11T01:46:20.720-04:00 _level=INFO _msg=Got new configuration for bucket default 
_time=2016-05-11T01:46:20.720-04:00 _level=INFO _msg=Got new configuration for bucket default 
_time=2016-05-11T01:46:20.720-04:00 _level=INFO _msg=Got new configuration for bucket default 
_time=2016-05-11T01:46:20.720-04:00 _level=INFO _msg=Got new configuration for bucket default 
_time=2016-05-11T01:46:20.720-04:00 _level=INFO _msg=Got new configuration for bucket default 
_time=2016-05-11T01:46:20.720-04:00 _level=INFO _msg=Got new configuration for bucket default 
_time=2016-05-11T01:46:20.720-04:00 _level=INFO _msg=Got new configuration for bucket default 
_time=2016-05-11T01:46:20.720-04:00 _level=INFO _msg=Got new configuration for bucket default 
_time=2016-05-11T01:46:20.720-04:00 _level=INFO _msg=Got new configuration for bucket default 
_time=2016-05-11T01:46:20.720-04:00 _level=INFO _msg=Got new configuration for bucket default 
_time=2016-05-11T01:46:20.723-04:00 _level=INFO _msg=Got new configuration for bucket default 
_time=2016-05-11T01:46:20.723-04:00 _level=INFO _msg=Got new configuration for bucket default 
_time=2016-05-11T01:46:20.724-04:00 _level=INFO _msg=Got new configuration for bucket default 
_time=2016-05-11T01:46:20.728-04:00 _level=INFO _msg=Got new configuration for bucket default

and the fact that view indexes are refreshed 1100 times per second.
There's loads of comments in the new code, but the issue goes as follows.

YCSB fires tons of queries at the same time on the same keyspace.
Keyspace not in the namespace cache, so in KeyspaceByName() all threads
don't find it and decide to to load it at the same time.
They all enter newKeyspace() which calls loadIndexes() (generates the load
on ddocs) and then RunBucketUpdater() (nice goroutine leak plus all the
"got new configuration..." log entries).

The new code serialises creation of the new keyspace through a keyspace
lock.
Once the thread gets the lock - it will check for the keyspace again, and
that will remove the load.
There will be a (hopefully small) spike on the newly introduced keyspace
lock (the good locks for this problem are patented by IBM), but after that
everything will be fine.

Also - logging the error, as it might be useful, but making sure that
we don't flood the log.

Change-Id: I86d228786f6bc8b39c78019f56a36e8071984a28
Reviewed-on: http://review.couchbase.org/75682
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>