MB-23610 delete by node does not work in system keyspaces
The uderlying issue here is that the delete is covered, but so far system
keyspaces indexes have only had keys but not values.
To make this work, we have to include the relevant node name as the
entry.

Also checked the the primary index - but for meta().id adding the value
to the index entry is not necessary.

Change-Id: Ibaaa74aace20dfdf5532de9a7e7eadcfc01a6d08
Reviewed-on: http://review.couchbase.org/76056
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>