MB-22467 use statement type from algebra for accounting purposes
so far accounting tokenized the statement and looked for tokens at specific
positions to determine what statement was being executed.
This was leading to wrong statements being tracked (think FROM...SELECT
but also prepared statements or explain).
This is - this information is readily available in the algebra.
Augmented the algebra to report the statement type, and added this
to prepared statements and requests.

Have done enough work to satisfy accounting needs - we can expad later.

Change-Id: I4b145be11c04d616c776806d9f70b556562d755f
Reviewed-on: http://review.couchbase.org/76073
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Gerald Sangudi <gerald@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>