MB-22657 remote monitoring via QWB MB-22709 n1ql clustering improvements
This Change picks up Dave F's fix for MB-22657 and augments it with parts
of what was intended for the clustering improvemens planned for MB-22709.
This is only intended to add enough improvements to have monitoring cope
with possibly changing node names, non rebalanced nodes and the like.
It does not yet deal with tracking down nodes and issuing warnings
when skipping nodes.
It also adds on top of Dave's code enough checks for n1ql services to 
be able to detect that they are not part of a cluster and therefore not
to meddle with it.

As Dave said in his review "this is work in progress"

Dave's original code: http://review.couchbase.org/#/c/76414/

Change-Id: Id1ec660f7d6f8c10f715d83e4c3eb0d0442aad4b
Reviewed-on: http://review.couchbase.org/76794
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>