MB-22708 datastore Count Fetch and Delete to forward warnings
The easiest thing here is to use QueryContex introduced by Johan, much
like the the indexer context already does.

Also - added an IsWarning() errors method and switched a ds_system
warning to having a WARNING level, so that we could switch
execution/fetch.go/flushBatch() or 
execution/delete_send.go/flushBatch() to deal with warnings from 
the datastore in a similar way as they deal with fatal errors, if we
so wished.
Change-Id: Id28503bfd34f7c4339a86b339baf967fe91d4dc5
Reviewed-on: http://review.couchbase.org/77377
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>