MB-24334 scanner throws panic on error
so far the scanner has used nex's own Error method to flag errors, however
this throws a straight panic, which essentially ends up with a stack on the
log.

Error cannot be overridden in nex, so we have to have our own.
Have added a functional pointer to the nex structure so that the scanner
can throw our own error.
Further - we can't use lex.Error, because that's used by the parser.
If we did, the net result would be two errors, one from the scanner, and
then another from the parser detecting the scanner error.
Trust me, it looks ugly.
Instead I have added a ScannerError() method which is passed to nex,
which reords the scanner error.
Error() will append it to the parser error, so the net result looks like this

    "requestID": "c549c63c-62f0-4eb9-a04d-64a478238671",
    "errors": [
        {
            "code": 3000,
            "msg": "syntax error: invalid quoted string - at 'q=select * from yahoo.finance.quotes where symbol in (\\\"HDP\\\") AND YearLow=\"6.42\"'"
        }
    ],
    
(nice), rather than 

    "requestID": "58e225fc-b544-46f5-9899-d6fc2dc6d852",
    "errors": [
        {
            "code": 3000,
            "msg": "invalid quoted string - at 'q=select * from yahoo.finance.quotes where symbol in (\\\"HDP\\\") AND YearLow=\"6.42\"' \n syntax error - at 'q=select * from yahoo.finance.quotes where symbol in (\\\"HDP\\\") AND YearLow=\"6.42\"'"
        }
    ],

(ugh!)

Change-Id: I7debeaee94be7969943c59fc83543452458e42be
Reviewed-on: http://review.couchbase.org/78128
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>