MB-18290 prepareds cache configurable and LRU cleaned
The bulk of this change is straight forward - use util/cache.go
rather than the exist ad hoc prepared cache and add configuration knobs.

The major change is in cache.go - essentially since we now move
statements to the top of the bucket at every execute, scans of the cache
would either skip lots of entries because the entry being scanned had
been moved to the head mid scan, or report duplicate entries, because an
entry already reported would be moved to the head before the scan
finished.

To avoid this - the cache now uses two lists - one for LRU cleanup and
one for scans.
There is also a new Use() method to mark an entry as MRU, and the
Add() method has been augmented to give the options to both mark as MRU
and decide if amend or not...

There's also a cache test.

Change-Id: I2f072f36e73eecad2e8872abebf313dfac8fec21
Reviewed-on: http://review.couchbase.org/78853
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>